back-end/requirements.txt
numpy==1.20.2
pandas==1.2.3
psycopg2-binary==2.8.4
pyodbc==4.0.30
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
